PyOcr/requirements.txt
numpy==1.18.1
opencv-python==4.1.0.25
Pillow==7.0.0
pyocr==0.7.2
